Manage the local state of the CLI wallet when using RPC
