+++ wallet/wallet-cli-lib/Cargo.toml
@@ -26,6 +26,7 @@ wallet-rpc-lib = { path = "../wallet-rpc-lib" }
 wallet-rpc-client = { path = "../wallet-rpc-client" }
 
 clap = { workspace = true, features = ["derive"] }
+async-trait.workspace = true
 crossterm.workspace = true
 directories.workspace = true
 humantime.workspace = true
